fix(experimental): physically move CodeEditor out of src/components
The docs sidebar generator (docs/.vitepress/utils.ts) categorizes purely by
folder location: anything under src/components/* with stories + a .md gets
a Components sidebar entry, anything under the top-level experimental/*
gets an Experimental one. CodeEditor kept its export re-pointed to
experimental.ts but stayed physically under src/components, so it still
showed up under Components even though the export moved.

Move experimental/CodeEditor to mirror Accordion/FloatingWindow/
MultiEmailInput exactly, fix the relative imports that changed depth
(InputLabeling components, useInputLabeling, inputTypes), and repoint
experimental.ts at the new path. This also gets CodeEditor a real
/docs/experimental/codeeditor page via the existing colocated-docs
mechanism, so add a short entry to the experimental overview pointing at it.
