+++ experimental.ts
@@ -11,10 +11,10 @@ export {
 } from './src/components/InputLabeling'
 export { useInputLabeling } from './src/composables/useInputLabeling'
 export type { FrappeUIError } from './src/composables/useInputLabeling'
-export { CodeEditor, CodePreview, loadLanguage } from './src/components/CodeEditor'
+export { CodeEditor, CodePreview, loadLanguage } from './experimental/CodeEditor'
 export type {
   CodeLanguage,
   CodeEditorProps,
   CodeEditorEmits,
   CodePreviewProps,
-} from './src/components/CodeEditor'
+} from './experimental/CodeEditor'
